requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-telegram==0.19.0
requests==2.32.5
telegram_text==0.2.0
urllib3==2.6.2
